DemoThree/requirements.txt
absl-py==0.9.0
altgraph==0.17
astor==0.8.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
future==0.18.2
gast==0.2.2
google-auth==1.11.2
google-auth-oauthlib==0.4.1
grpcio==1.27.2
h5py==2.10.0
idna==2.9
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2.1
matplotlib==2.1.0
mock==4.0.2
numpy==1.18.1
oauthlib==3.1.0
olefile==0.46
opencv-python==4.2.0.32
pefile==2019.4.18
Pillow==4.2.1
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyinstaller==4.0.dev0+96897fc78d
pyparsing==2.4.6
PyQt5==5.13.0
PyQt5-sip==12.7.1
pyqt5-tools==5.13.0.1.5
python-dateutil==2.8.1
python-dotenv==0.11.0
pytz==2019.3
pywin32-ctypes==0.2.0
PyYAML==5.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.14.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
urllib3==1.25.8
Werkzeug==1.0.0
